+++ internal/pkg/server_manager/events/events.go
@@ -3,6 +3,7 @@ package events
 import (
 	"github.com/assetto-corsa-web/accweb/internal/pkg/server_manager"
 	"github.com/assetto-corsa-web/accweb/internal/pkg/server_manager/events/callback"
+	"github.com/assetto-corsa-web/accweb/internal/pkg/server_manager/events/globalentrylist"
 	"github.com/assetto-corsa-web/accweb/internal/pkg/server_manager/events/logparser"
 	"github.com/assetto-corsa-web/accweb/internal/pkg/server_manager/events/logwriter"
 	"github.com/assetto-corsa-web/accweb/internal/pkg/server_manager/events/windowsadv"
@@ -13,4 +14,5 @@ func InitializeAll(sm *server_manager.Service) {
 	logparser.Register(sm)
 	windowsadv.Register(sm)
 	callback.Register(sm)
+	globalentrylist.Register(sm)
 }