+++ public/src/pages/configurations/index.js
@@ -1,0 +1,2 @@
+export { default as globalentrylist } from "./globalentrylist.vue";
+export { default as globalbanlist } from "./globalbanlist.vue";