+++ public/src/store/store.js
@@ -1,6 +1,6 @@
 import Vuex from "vuex";
-import {AuthStore} from "./auth.js";
-import {ErrorStore} from "./error";
+import { AuthStore } from "./auth.js";
+import { ErrorStore } from "./error";
 
 export default function NewStore() {
 	return new Vuex.Store({